pyproject.toml
[project]
name = "english-practice-api"
version = "0.1.0"
description = "Backend API for English Practice web application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
]

[project.scripts]
start = "uvicorn api.main:app --reload --port 8000"

[dependency-groups]
dev = []
